Processing.js now integrated
Assumed place for *.pde is http://folk.ntnu.no/[USERNAME]/plab/plab.pde
ONLY this pde can be loaded at this moment.
Bridge to javascript updated. setSize() caused issues with rezising of
canvas.
A security issue may be that the pde is not checked in any way before
load. However a student should be able to control the pde he or she
visits.
Next step is to continue with bluetooth integration