Loading of processing may now be cancelled
A countdown shows how many attempts are made.
New actual attempts are now done. Checks if they were successful are
performed.
There is still an issue concerning connections when using wifi. They do
not always lead to a result. The reason is unknown, and seen as a area for
further study.