English and Norwegian supported. Fix #2
All text that should be translated should at this point be correctly
translated. Translation of module names not added, they should be language
independent. However, this may change in the future.

Styles for language buttons are not completed. Should be improved upon
creating settings screen.

Dom elements for settings tab are added.