International and settings. Fix #5, fix #11
It is now possible to add setting items in a consistent way.
All functionality (e.g. images, single-select) has not been tested yet.
It is also possible to access any server from the app. A setting is added
so the base location and a postfix can be edited by the user of the app.
The final location still use the username that is requested. The text of
this should be improved, as it is no longer username. An issue will be
created on GitHub to track this.