BTLE service can be set, fix #7
Improvement might be to allow user to set both serviceuuid, and rx/tx
characteristic uuids. Two services supported, the one used by Nordic
Semiconductor in their test app (which also Adafruit use on their BTLE
units and libraries) and the one that was supported before (strange
service with uuid FFE0)

Also removed text on back button that is visible while sketch is running.
The back arrow is seen as enough to know what it means, and give less
cluttered view.