Fix #19
It is now possible to select multiple .pde files to load. These will all
follow the same pattern as the main .pde file when it come to url
location. It is intended to use if a project has multiple .pde files in
same directory. The addresses for the selected .pde files are not stored
between each session.
Multiple locations (url base and postfix for file(s)) may be further
specified in settings between adding the different files. Url string is
built when pressing "ok".